Update dependencies and fix NDArray shape casting
Upgraded multiple NuGet packages to latest versions in both PixelsorterApp and PixelsorterClassLib. Explicitly cast NDArray shape values to int for improved type safety and compatibility with updated NumSharp.

+++ PixelsorterClassLib/Core/Image.cs
@@ -59,9 +59,9 @@ public class Image
     public static Image<Rgba32> NdarrayToImgData(NDArray data)
     {
         var shape = data.shape;
-        int height = shape[0];
-        int width = shape[1];
-        int channels = shape[2];
+        int height = (int)shape[0];
+        int width = (int)shape[1];
+        int channels = (int)shape[2];
 
         var sourceData = data.ToArray<float>();
